splitted handler files and created setup when starting either back or front-end
