+++ back-end/src/sciler/handler/handler_test.go
@@ -249,739 +249,6 @@ func TestOnConnectionMsg(t *testing.T) {
 	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
 }
 
-////////////////////////////// Status tests //////////////////////////////
-func TestOnStatusMsg(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"testComponent0": true,
-			"testComponent1": 40,
-			"testComponent2": "blue",
-			"testComponent3": []int{1, 2, 3},
-			"testComponent4": true,
-			"testComponent5": 40,
-			"testComponent6": "blue",
-			"testComponent7": []int{1, 2, 3},
-			"testComponent8": "custom",
-			"testComponent9": "custom",
-		},
-	}
-	handler.updateStatus(msg)
-
-	tests := []struct {
-		name      string
-		component string
-		status    interface{}
-	}{
-		{
-			name:      "component0 test",
-			component: "testComponent0",
-			status:    true,
-		},
-		{
-			name:      "component1 test",
-			component: "testComponent1",
-			status:    40,
-		},
-		{
-			name:      "component2 test",
-			component: "testComponent2",
-			status:    "blue",
-		},
-		{
-			name:      "component3 test",
-			component: "testComponent3",
-			status:    []int{1, 2, 3},
-		},
-		{
-			name:      "component4 test",
-			component: "testComponent4",
-			status:    true,
-		},
-		{
-			name:      "component5 test",
-			component: "testComponent5",
-			status:    40,
-		},
-		{
-			name:      "component6 test",
-			component: "testComponent6",
-			status:    "blue",
-		},
-		{
-			name:      "component7 test",
-			component: "testComponent7",
-			status:    []int{1, 2, 3},
-		},
-		// TODO implement custom
-		{
-			name:      "component8 test",
-			component: "testComponent8",
-			status:    nil,
-		},
-		{
-			name:      "component9 test",
-			component: "testComponent9",
-			status:    nil,
-		}}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.status, handler.Config.Devices["TestDevice"].Status[tt.component],
-				"Device should set status right")
-		})
-	}
-}
-
-func TestOnStatusMsgOtherDevice(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "WrongDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"testComponent0": false,
-			"testComponent1": true,
-			"testComponent2": false},
-	}
-	handler.updateStatus(msg)
-
-	_, ok := handler.Config.Devices["WrongDevice"]
-	assert.Equal(t, false, ok,
-		"Device should not exist in devices because it was not in config")
-}
-
-func TestOnStatusMsgWrongComponent(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"wrongComponent": true},
-	}
-	handler.updateStatus(msg)
-
-	_, ok := handler.Config.Devices["TestDevice"].Status["wrongComponent"]
-	assert.Equal(t, false, ok,
-		"Component should not exist in device because it was not in config")
-}
-
-func TestOnStatusMsgWrongType(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"testComponent1": true,
-			"testComponent2": 40,
-			"testComponent3": "blue",
-			"testComponent0": []int{1, 2, 3},
-			"testComponent5": true,
-			"testComponent6": 40,
-			"testComponent7": "blue",
-			"testComponent4": []int{1, 2, 3}},
-	}
-	handler.updateStatus(msg)
-
-	tests := []struct {
-		name      string
-		component string
-	}{
-		{
-			name:      "component0 test",
-			component: "testComponent0",
-		},
-		{
-			name:      "component1 test",
-			component: "testComponent1",
-		},
-		{
-			name:      "component2 test",
-			component: "testComponent2",
-		},
-		{
-			name:      "component3 test",
-			component: "testComponent3",
-		},
-		{
-			name:      "component4 test",
-			component: "testComponent4",
-		},
-		{
-			name:      "component5 test",
-			component: "testComponent5",
-		},
-		{
-			name:      "component6 test",
-			component: "testComponent6",
-		},
-		{
-			name:      "component7 test",
-			component: "testComponent7",
-		}}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, ok := handler.Config.Devices["TestDevice"].Status[tt.component]
-			assert.Equal(t, false, ok,
-				"component should not been updated in device because it was not the right type")
-		})
-	}
-}
-
-func TestMsgMapperStatus(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"testComponent0": true,
-			"testComponent1": 40,
-			"testComponent2": "blue"},
-	}
-	handler.msgMapper(msg)
-
-	assert.Equal(t, true, handler.Config.Devices["TestDevice"].Status["testComponent0"],
-		"Device should set status to true on component 0")
-	assert.Equal(t, 40, handler.Config.Devices["TestDevice"].Status["testComponent1"],
-		"Device should set status to 40 on component 1")
-	assert.Equal(t, "blue", handler.Config.Devices["TestDevice"].Status["testComponent2"],
-		"Device should set status to blue on component 2")
-}
-
-////////////////////////////// Confirmation tests //////////////////////////////
-func TestOnConfirmationMsgFrontEnd(t *testing.T) {
-	communicatorMock := new(CommunicatorMock)
-	workingConfig := config.ReadFile("../../../resources/testing/test_config.json")
-	handler := Handler{
-		Config:       workingConfig,
-		ConfigFile:   "../../../resources/testing/test_config.json",
-		Communicator: communicatorMock,
-	}
-	msg := Message{
-		DeviceID: "controlBoard",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": true,
-			"instructed": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"type":      "instruction",
-				"contents": []map[string]interface{}{{
-					"instruction":   "test",
-					"instructed_by": "front-end"},
-				},
-			},
-		},
-	}
-	jsonMsg, _ := json.Marshal(&msg)
-	communicatorMock.On("Publish", "front-end", string(jsonMsg), 3)
-	handler.onConfirmationMsg(msg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestOnConfirmationMsgTrue(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": true,
-			"instructed": map[string]interface{}{
-				"device_id": "front-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"contents":  []map[string]interface{}{{"instruction": "test"}},
-				"type":      "instruction",
-			},
-		},
-	}
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, true, handler.Config.Devices["TestDevice"].Connection,
-		"Device's connection should have been set to true")
-}
-
-func TestOnConfirmationMsgFalse(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": false,
-			"instructed": map[string]interface{}{
-				"device_id": "front-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"contents": []map[string]interface{}{{
-					"instruction":   "test",
-					"instructed_by": "front-end"}},
-				"type": "instruction",
-			},
-		},
-	}
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, true, handler.Config.Devices["TestDevice"].Connection,
-		"Device's connection should have been set to true")
-}
-
-func TestOnConfirmationMsgIncorrect1(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completion": true,
-			"instructed": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"contents": []map[string]interface{}{{
-					"instruction":   "test",
-					"instructed_by": "front-end"}},
-				"type": "instruction",
-			},
-		},
-	}
-	before := handler.Config
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, before, handler.Config,
-		"Device should not alter config upon invalid confirmation message with no completed value")
-}
-
-func TestOnConfirmationMsgIncorrect2(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": true,
-			"instruction": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"type":      "instruction",
-				"contents": map[string]interface{}{
-					"instructions":  "test",
-					"instructed_by": "front-end"},
-			},
-		},
-	}
-	before := handler.Config
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, before, handler.Config,
-		"Device should not alter config upon invalid confirmation message with no instructed object")
-}
-
-func TestOnConfirmationMsgIncorrect3(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": "true",
-			"instructed": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"contents": []map[string]interface{}{{
-					"instruction":   "test",
-					"instructed_by": "front-end"}},
-				"type": "instruction",
-			},
-		},
-	}
-	before := handler.Config
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, before, handler.Config,
-		"Device should not alter config upon invalid confirmation message with no boolean completed")
-}
-
-func TestOnConfirmationMsgIncorrect4(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": true,
-			"instructed": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"type":      "instruction",
-				"contents": map[string]interface{}{
-					"instruction": "test",
-				},
-			},
-		},
-	}
-	before := handler.Config
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, before, handler.Config,
-		"Device should not alter config upon invalid confirmation message with no instructions list")
-}
-
-func TestOnConfirmationMsgIncorrect5(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "display",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": true,
-			"instructed": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"type":      "instruction",
-				"contents": []map[string]interface{}{
-					{"instruction": "test"},
-				},
-			},
-		},
-	}
-	before := handler.Config
-	handler.onConfirmationMsg(msg)
-	assert.Equal(t, before, handler.Config,
-		"Device should not alter config upon invalid confirmation message with device id not in config")
-}
-
-func TestMsgMapperConfirmation(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "confirmation",
-		Contents: map[string]interface{}{
-			"completed": true,
-			"instructed": map[string]interface{}{
-				"device_id": "back-end",
-				"time_sent": "05-12-2019 09:42:10",
-				"contents": []map[string]interface{}{{
-					"instruction":   "test",
-					"instructed_by": "front-end"}},
-				"type": "instruction",
-			},
-		},
-	}
-	before := handler.Config
-	handler.msgMapper(msg)
-	assert.Equal(t, before, handler.Config,
-		"Device should not alter config upon confirmation message")
-}
-
-////////////////////////////// Instruction tests //////////////////////////////
-func TestOnInstructionMsgName(t *testing.T) {
-	msg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{
-			{"instruction": "send name"},
-		},
-	}
-	communicatorMock := new(CommunicatorMock)
-	handler := Handler{
-		Config:       config.ReadFile("../../../resources/testing/test_instruction.json"),
-		Communicator: communicatorMock,
-	}
-	returnMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "name",
-		Contents: map[string]interface{}{
-			"name": "Escape X",
-		},
-	})
-	communicatorMock.On("Publish", "front-end", string(returnMessage), 3)
-	handler.onInstructionMsg(msg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestOnInstructionMsgTestAll(t *testing.T) {
-	communicatorMock := new(CommunicatorMock)
-	workingConfig := config.ReadFile("../../../resources/testing/test_config.json")
-	handler := Handler{
-		Config:       workingConfig,
-		Communicator: communicatorMock,
-	}
-	instructionMsg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{
-			{"instruction": "test all"},
-		},
-	}
-	responseMsg, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction":   "test",
-			"instructed_by": "front-end"},
-		},
-	})
-	communicatorMock.On("Publish", "client-computers", string(string(responseMsg)), 3)
-	handler.onInstructionMsg(instructionMsg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestOnInstructionMsgSendStatus(t *testing.T) {
-	msg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{
-			{"instruction": "send status"},
-		},
-	}
-	communicatorMock := new(CommunicatorMock)
-	handler := Handler{
-		Config:       config.ReadFile("../../../resources/testing/test_instruction_status.json"),
-		Communicator: communicatorMock,
-	}
-	statusMessageDisplay, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"id": "display", "connection": false, "status": map[string]interface{}{},
-		},
-	})
-	statusMessageFrontEnd, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "status",
-		Contents: map[string]interface{}{
-			"id": "front-end", "connection": false, "status": map[string]interface{}{"start": 0, "stop": 0},
-		},
-	})
-	timerMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "time",
-		Contents: map[string]interface{}{
-			"state":    "stateIdle",
-			"duration": 10000,
-			"id":       "timer1",
-		},
-	})
-	timerGeneralMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "time",
-		Contents: map[string]interface{}{
-			"state":    "stateIdle",
-			"duration": 60000,
-			"id":       "general",
-		},
-	})
-	eventMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "event status",
-		Contents: []map[string]interface{}{
-			{"id": "rule", "status": true, "description": "My rule"},
-		},
-	})
-	communicatorMock.On("Publish", "front-end", string(eventMessage), 3)
-	communicatorMock.On("Publish", "front-end", string(timerMessage), 3)
-	communicatorMock.On("Publish", "front-end", string(timerGeneralMessage), 3)
-	communicatorMock.On("Publish", "front-end", string(statusMessageDisplay), 3)
-	communicatorMock.On("Publish", "front-end", string(statusMessageFrontEnd), 3)
-	handler.onInstructionMsg(msg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 5)
-}
-
-func TestOnInstructionMsgResetAll(t *testing.T) {
-	communicatorMock := new(CommunicatorMock)
-	workingConfig := config.ReadFile("../../../resources/testing/test_config.json")
-	handler := Handler{
-		Config:       workingConfig,
-		ConfigFile:   "../../../resources/testing/test_config.json",
-		Communicator: communicatorMock,
-	}
-	instructionMsg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{
-			{"instruction": "reset all"},
-		},
-	}
-	responseMsg := Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction":   "reset",
-			"instructed_by": "front-end"},
-		},
-	}
-	statusMsg := Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "time",
-		Contents: map[string]interface{}{
-			"id":       "general",
-			"duration": 1800000,
-			"state":    "stateIdle",
-		},
-	}
-
-	jsonInstructionMsg, _ := json.Marshal(&responseMsg)
-	jsonStatusMsg, _ := json.Marshal(&statusMsg)
-
-	communicatorMock.On("Publish", "client-computers", string(jsonInstructionMsg), 3)
-	communicatorMock.On("Publish", "front-end", string(jsonInstructionMsg), 3)
-	communicatorMock.On("Publish", "front-end", string(jsonStatusMsg), 3)
-	handler.msgMapper(instructionMsg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 3)
-}
-
-func TestOnInstructionMsgHint(t *testing.T) {
-	msg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction": "hint",
-			"value":       "This is my hint"},
-		},
-	}
-	communicatorMock := new(CommunicatorMock)
-	handler := Handler{
-		Config:       config.ReadFile("../../../resources/testing/test_instruction.json"),
-		Communicator: communicatorMock,
-	}
-	returnMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction":   "hint",
-			"value":         "This is my hint",
-			"instructed_by": "front-end"},
-		},
-	})
-	communicatorMock.On("Publish", "hint", string(returnMessage), 3)
-	handler.onInstructionMsg(msg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestInstructionAllHints(t *testing.T) {
-	communicatorMock := new(CommunicatorMock)
-	workingConfig := config.ReadFile("../../../resources/testing/test_config.json")
-	handler := Handler{
-		Config:       workingConfig,
-		ConfigFile:   "../../../resources/testing/test_config.json",
-		Communicator: communicatorMock,
-	}
-	instructionMsg := Message{
-		DeviceID: "front-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "instruction",
-		Contents: []map[string]interface{}{
-			{
-				"instruction": "all hints",
-			},
-		},
-	}
-	returnMessage := Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "all hints",
-		Contents: map[string][]string{
-			"Telefoon puzzels": {"De knop verzend jouw volgorde", "Heb je al even gewacht?"},
-			"Control puzzel":   {"Zet de schuiven nauwkeurig"},
-		},
-	}
-	jsonHintMessage, _ := json.Marshal(&returnMessage)
-	communicatorMock.On("Publish", "front-end", string(jsonHintMessage), 3)
-	handler.msgMapper(instructionMsg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestOnInstructionMsgFinishRule(t *testing.T) {
-	msg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction": "finish rule",
-			"rule":        "rule"},
-		},
-	}
-	communicatorMock := new(CommunicatorMock)
-	handler := Handler{
-		Config:       config.ReadFile("../../../resources/testing/test_instruction.json"),
-		Communicator: communicatorMock,
-	}
-	returnMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "event status",
-		Contents: []map[string]interface{}{
-			{"id": "rule", "description": "My rule", "status": true},
-		},
-	})
-	communicatorMock.On("Publish", "front-end", string(returnMessage), 3)
-	handler.onInstructionMsg(msg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestOnInstructionMsgTestDevice(t *testing.T) {
-	msg := Message{
-		DeviceID: "front-end",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction": "test device",
-			"device":      "display"},
-		},
-	}
-	communicatorMock := new(CommunicatorMock)
-	handler := Handler{
-		Config:       config.ReadFile("../../../resources/testing/test_instruction.json"),
-		Communicator: communicatorMock,
-	}
-	returnMessage, _ := json.Marshal(Message{
-		DeviceID: "back-end",
-		TimeSent: time.Now().Format("02-01-2006 15:04:05"),
-		Type:     "instruction",
-		Contents: []map[string]interface{}{{
-			"instruction":   "test",
-			"instructed_by": "front-end"},
-		},
-	})
-	communicatorMock.On("Publish", "display", string(returnMessage), 3)
-	handler.onInstructionMsg(msg)
-	communicatorMock.AssertNumberOfCalls(t, "Publish", 1)
-}
-
-func TestOnInstructionMsgMapper(t *testing.T) {
-	handler := getTestHandler()
-	msg := Message{
-		DeviceID: "TestDevice",
-		TimeSent: "05-12-2019 09:42:10",
-		Type:     "instruction",
-		Contents: []map[string]interface{}{
-			{"instruction": "name"},
-		},
-	}
-
-	before := handler.Config
-	handler.msgMapper(msg)
-	assert.Equal(t, before, handler.Config,
-		"Nothing should have been changed after an instruction message type")
-}
-
 ////////////////////////////// Event handling tests //////////////////////////////
 func TestHandleSingleEvent(t *testing.T) {
 	communicatorMock := new(CommunicatorMock)
