+++ back-end/src/sciler/main.go
@@ -39,11 +39,7 @@ func main() {
 
 	messageHandler := handler.Handler{Config: configurations, ConfigFile: filename, Communicator: communicator}
 	go communicator.Start(messageHandler.NewHandler, func() {
-		messageHandler.SendStatus("general")
-		for _, value := range messageHandler.Config.Devices {
-			messageHandler.SendStatus(value.ID)
-			messageHandler.GetStatus(value.ID)
-		}
+		messageHandler.SendSetUp()
 	})
 
 	// prevent exit
